==> markdown_files/Transcriptomic/Pipeline_Scripts/1B_STAR_pipeline/outdated/STAR_pipeline_withquantModeCounts/03C_CV17_RNA_DiffExpression.Rmd <==
---
title: "03 STAR DiffExpression"
author: "adowneywall"
date: "6/11/2019"
output: html_document
editor_options: 
  chunk_output_type: console
---

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
library(DESeq2)
library(ashr)
library(kableExtra)
```


### **Data**  
  
```{r}
model<-read.csv("/home/downeyam/Github/2017OAExp_Oysters/input_files/RNA/metadata_cvirginica_rna_meta.txt", header=TRUE)
model$Treatment <- as.factor(model$treatment)
model$Time <- as.character(model$timepoint)
model$Time[model$Time == "3"] <- "09"
model$Time[model$Time == "6"] <- "80"
model$Time <- as.factor(model$Time)
model$Pop <- as.factor(model$population)
model$Lane <- as.factor(model$lane)
model$SFV <-  interaction(model$Time,model$Treatment) # Creates single factor variable for combination of time and treatment

# Raw counts based on --quantMode
gc_unstranded <- readRDS("/home/downeyam/Github/2017OAExp_Oysters/input_files/RNA/STAR_pipeline/rawCounts/run20190610_GeneCountMatrix_unstranded.rds")
colnames(gc_unstranded) <- model$sample_name
gc_stranded <- readRDS("/home/downeyam/Github/2017OAExp_Oysters/input_files/RNA/STAR_pipeline/rawCounts/run20190610_GeneCountMatrix_Stranded.rds")
colnames(gc_stranded) <- model$sample_name
gc_reverse <- readRDS("/home/downeyam/Github/2017OAExp_Oysters/input_files/RNA/STAR_pipeline/rawCounts/run20190610_GeneCountMatrix_Reverse.rds")
colnames(gc_reverse) <- model$sample_name

# Additional info about each gene
trans <- readRDS("/home/downeyam/Github/2017OAExp_Oysters/input_files/RNA/transcriptome_table.RData")
trans_genes <- trans[match(regmatches(trans$transVariant,regexpr(">.*",trans$transVariant)),trans$transVariant),]
```
  
### DEGs based on unstranded GeneCounts  
  
**Full Model w/ unstranded Counts** - Lane, Pop, Treatment, Time, Treatment:Time as factors  
```{r}
# Full Model
full <- DESeqDataSetFromMatrix(countData = gc_unstranded,
                              colData = model,
                              design = ~ Lane + Pop + Treatment + Time + Treatment:Time)

dds_full_Wald <- DESeq(full,minmu = 80,test = "Wald")
# Interaction
res_full <- results(dds_full_Wald)
?results()
summary(res_full)
pVal_interaction <- res_full$padj
# Time
res_full_Time <- results(dds_full_Wald,contrast=c("Time","09","80"))
summary(res_full_Time)
pVal_time <- res_full_Time$padj
# Treatment
res_full_Treatment <- results(dds_full_Wald,contrast=c("Treatment","400","2800"))
summary(res_full_Treatment)
pVal_treatment <- res_full_Treatment$padj

pvals_df <- data.frame(location=row.names(gc_unstranded),pVal_interaction,pVal_time,pVal_treatment)

gene_GE <- merge(trans_genes,pvals_df,by="location")
```
  
**Top Twenty Genes Associated with Treatment**  
```{r}
gene_GE_trt <- gene_GE[order(gene_GE$pVal_treatment),]
kable(gene_GE_trt[1:20,]) %>%
  kable_styling()
```
  
**Top Twenty Genes Associated with Time**  
```{r}
gene_GE_t <- gene_GE[order(gene_GE$pVal_time),]
kable(gene_GE_t[1:20,]) %>%
  kable_styling()
```
  
### DEGs based on stranded (Read 1) GeneCounts  
  
**Full Model w/ stranded Counts** - Lane, Pop, Treatment, Time, Treatment:Time as factors  
```{r}
# Full Model
full <- DESeqDataSetFromMatrix(countData = gc_stranded,
                              colData = model,
                              design = ~ Lane + Pop + Treatment + Time + Treatment:Time)

dds_full_Wald <- DESeq(full,minmu = 80,test = "Wald")
# Interaction
res_full <- results(dds_full_Wald)
summary(res_full)
pVal_interaction <- res_full$padj
# Time
res_full_Time <- results(dds_full_Wald,contrast=c("Time","09","80"))
summary(res_full_Time)
pVal_time <- res_full_Time$padj
# Treatment
res_full_Treatment <- results(dds_full_Wald,contrast=c("Treatment","400","2800"))
summary(res_full_Treatment)
pVal_treatment <- res_full_Treatment$padj

pvals_df <- data.frame(location=row.names(gc_unstranded),pVal_interaction,pVal_time,pVal_treatment)

gene_GE <- merge(trans_genes,pvals_df,by="location")
```
  
**Top Twenty Genes Associated with Treatment**  
```{r}
gene_GE_trt <- gene_GE[order(gene_GE$pVal_treatment),]
kable(gene_GE_trt[1:20,]) %>%
  kable_styling()
```
  
**Top Twenty Genes Associated with Time**  
```{r}
gene_GE_t <- gene_GE[order(gene_GE$pVal_time),]
kable(gene_GE_t[1:20,]) %>%
  kable_styling()
```
